requirements.txt
numpy
pandas
matplotlib
seaborn
python-dotenv
psycopg2
SQLAlchemy
nltk
streamlit